refactor(cli): flatten workflow command module
Move the Workflow command slice to the neutral command root and rename its internal export without changing public CLI, token, webhook, or API compatibility behavior.
