+++ turbo/apps/cli/src/commands/workflow/__tests__/delete.test.ts
@@ -9,7 +9,7 @@
 
 import { describe, it, expect, vi, beforeEach, afterEach } from "vitest";
 import { http, HttpResponse } from "msw";
-import { server } from "../../../../mocks/server";
+import { server } from "../../../mocks/server";
 import { deleteCommand } from "../delete";
 import chalk from "chalk";
