+++ turbo/apps/cli/src/commands/workflow/__tests__/list.test.ts
@@ -9,7 +9,7 @@
 
 import { describe, it, expect, vi, beforeEach, afterEach } from "vitest";
 import { http, HttpResponse } from "msw";
-import { server } from "../../../../mocks/server";
+import { server } from "../../../mocks/server";
 import { listCommand } from "../list";
 import chalk from "chalk";
