+++ turbo/apps/cli/src/commands/workflow/__tests__/view.test.ts
@@ -9,7 +9,7 @@
 
 import { describe, it, expect, vi, beforeEach, afterEach } from "vitest";
 import { http, HttpResponse } from "msw";
-import { server } from "../../../../mocks/server";
+import { server } from "../../../mocks/server";
 import { viewCommand } from "../view";
 import chalk from "chalk";
