+++ turbo/apps/cli/src/commands/workflow/automation/display.ts
@@ -1,9 +1,9 @@
 import chalk from "chalk";
 import type { ChatThreadServiceTier } from "@okouai/api-contracts/contracts/chat-threads";
 import type { GmailNewMessageEventConfig } from "@okouai/api-contracts/contracts/zero-workflows";
-import type { ZeroWorkflowAutomationSummary } from "../../../../lib/api/domains/zero-workflows";
-import { formatRelativeTime } from "../../../../lib/domain/relative-time";
-import { formatDurationSeconds } from "../../../shared/duration";
+import type { ZeroWorkflowAutomationSummary } from "../../../lib/api/domains/zero-workflows";
+import { formatRelativeTime } from "../../../lib/domain/relative-time";
+import { formatDurationSeconds } from "../../shared/duration";
 
 type GmailMatchRules = NonNullable<GmailNewMessageEventConfig["match"]>;
 type GmailTextMatcher = NonNullable<GmailMatchRules["from"]>;
