+++ turbo/apps/cli/src/commands/workflow/automation/index.ts
@@ -28,12 +28,12 @@ import {
   listWorkspaceWorkflowAutomations,
   listWorkflowAutomations,
   updateWorkflowAutomation,
-} from "../../../../lib/api/domains/zero-workflows";
-import { getZeroChatThread } from "../../../../lib/api/domains/zero-chat";
-import { listZeroModelPolicies } from "../../../../lib/api/domains/zero-model-policies";
-import { withErrorHandler } from "../../../../lib/command/with-error-handler";
-import { decodeZeroTokenPayload } from "../../../../lib/api/zero-token";
-import { parseDurationSeconds } from "../../../shared/duration";
+} from "../../../lib/api/domains/zero-workflows";
+import { getZeroChatThread } from "../../../lib/api/domains/zero-chat";
+import { listZeroModelPolicies } from "../../../lib/api/domains/zero-model-policies";
+import { withErrorHandler } from "../../../lib/command/with-error-handler";
+import { decodeZeroTokenPayload } from "../../../lib/api/zero-token";
+import { parseDurationSeconds } from "../../shared/duration";
 import {
   resolveWorkflowRef,
   type WorkflowRefOptions,
