+++ turbo/apps/cli/src/commands/workflow/create.ts
@@ -1,10 +1,10 @@
 import { Command } from "commander";
 import chalk from "chalk";
 import { readFileSync } from "node:fs";
-import { createWorkflow } from "../../../lib/api/domains/zero-workflows";
-import { withErrorHandler } from "../../../lib/command/with-error-handler";
-import { readSupplementaryFiles } from "../../../lib/skill-directory";
-import { getOkouAgentId, getOkouChatThreadId } from "../../../lib/okou-env";
+import { createWorkflow } from "../../lib/api/domains/zero-workflows";
+import { withErrorHandler } from "../../lib/command/with-error-handler";
+import { readSupplementaryFiles } from "../../lib/skill-directory";
+import { getOkouAgentId, getOkouChatThreadId } from "../../lib/okou-env";
 import { formatWorkflowAgentName } from "./format";
 
 export const createCommand = new Command()
