+++ turbo/apps/cli/src/commands/workflow/delete.ts
@@ -3,9 +3,9 @@ import chalk from "chalk";
 import {
   deleteWorkflow,
   getWorkflow,
-} from "../../../lib/api/domains/zero-workflows";
-import { isInteractive, promptConfirm } from "../../../lib/utils/prompt-utils";
-import { withErrorHandler } from "../../../lib/command/with-error-handler";
+} from "../../lib/api/domains/zero-workflows";
+import { isInteractive, promptConfirm } from "../../lib/utils/prompt-utils";
+import { withErrorHandler } from "../../lib/command/with-error-handler";
 import {
   resolveWorkflowRef,
   type WorkflowRefOptions,
