+++ turbo/apps/cli/src/commands/workflow/edit.ts
@@ -1,9 +1,9 @@
 import { Command } from "commander";
 import chalk from "chalk";
 import { readFileSync } from "node:fs";
-import { updateWorkflow } from "../../../lib/api/domains/zero-workflows";
-import { withErrorHandler } from "../../../lib/command/with-error-handler";
-import { readSupplementaryFiles } from "../../../lib/skill-directory";
+import { updateWorkflow } from "../../lib/api/domains/zero-workflows";
+import { withErrorHandler } from "../../lib/command/with-error-handler";
+import { readSupplementaryFiles } from "../../lib/skill-directory";
 import {
   resolveWorkflowRef,
   type WorkflowRefOptions,
